perf: use int32 for IVF partition IDs to halve index RAM
Partition IDs for 20M vectors: int64 = 160 MB, int32 = 80 MB.
int32 is sufficient (max ~2.1B >> 20M). Also switch sorted_candidates
from candidates[argsort] to np.sort to preserve int32 dtype.

+++ scripts/build_index.py
@@ -119,11 +119,12 @@ def build_index(
     print(f"  Assignment done in {time.perf_counter() - t0:.1f}s")
 
     # --- Build CSR arrays ---
+    # int32 for IDs: max 2^31-1 > 20M, halves RAM vs int64
     counts = np.array([len(p) for p in partition_lists], dtype=np.int64)
     offsets = np.concatenate([[0], np.cumsum(counts)]).astype(np.int64)
 
-    non_empty = [np.array(p, dtype=np.int64) for p in partition_lists if p]
-    ids = np.concatenate(non_empty) if non_empty else np.array([], dtype=np.int64)
+    non_empty = [np.array(p, dtype=np.int32) for p in partition_lists if p]
+    ids = np.concatenate(non_empty) if non_empty else np.array([], dtype=np.int32)
 
     # --- Save ---
     index_dir.mkdir(parents=True, exist_ok=True)
